reworking calls. successful call to back-end and display for item
