+++ src/frontend/public/assets/search.png
