+++ src/frontend/public/assets/search_small.png
