+++ src/frontend/src/components/Results/ResultsComponent.tsx
@@ -3,7 +3,7 @@ import CraftingComponent from './CraftComponent';
 import PricesComponent from './PriceComponent';
 import ItemsComponent from './ItemComponent';
 import RecipesComponent from './RecipeComponent';
-import '../styles/Results.css';
+import '../../styles/Results.css';
 
 interface ResultsProps {
     componentName: string;
