+++ src/frontend/src/styles/SearchInputs.css
@@ -15,15 +15,19 @@
 }
 
 .search-input {
+    text-align: center;
     margin: 15px 5px 5px;
     font-size: 22px;
     width: 100px;
     padding: 5px 5px 5px 10px;
+    min-width: 100px;
 }
 
 .search-dropdown {
+    text-align: center;
     margin: 5px;
     font-size: 22px;
     width: 100px;
     padding: 5px 10px 5px 20px;
+    min-width: 120px;
 }
